requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anaconda==0.0.1.1
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
audioread==3.0.0
Automat==20.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==6.1.0
blessed==1.20.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
cloudpickle==2.2.1
cmake==3.26.3
colabcode==0.3.0
colorama==0.4.4
comm==0.2.2
configobj==5.0.6
constantly==15.1.0
contourpy @ file:///croot/contourpy_1700583582875/work
cycler==0.11.0
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1713537449725/work
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
discord==2.2.2
discord.py==2.3.0
distlib==0.3.8
efficientnet_pytorch==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.4
fonttools==4.29.1
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
fs==2.4.12
fsspec==2024.2.0
ftfy==6.2.0
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.42
glide-text2im @ git+https://github.com/openai/glide-text2im@69b530740eb6cef69442d6180579ef5ba9ef063e
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gpustat==1.1.1
greenlet==1.1.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperlink==21.0.0
idna==3.7
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio==2.34.0
incremental==21.3.0
install==1.3.5
ipykernel==6.29.4
ipython==8.24.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.3.2
labelImg==1.8.6
lazy_loader==0.3
librosa==0.10.0.post2
lit==16.0.3
llvmlite==0.40.0
lxml==4.9.2
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1706514640841/work
multitasking==0.0.11
munch==4.0.0
nbclassic==0.3.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
neat-python==0.92
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.1.0
notebook==7.1.3
notebook_shim==0.2.4
numba==0.57.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
olefile==0.46
opencv-python==4.9.0.80
openpyxl==3.1.2
outcome==1.2.0
overrides==7.7.0
packaging==24.0
pandas==2.0.0
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.1
ply==3.11
pooch==1.6.0
pretrainedmodels==0.7.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1700596371886/work
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.17
pycryptodomex==3.17
PyDrive==1.3.1
Pygments==2.17.2
PyHamcrest==2.0.2
PyMuPDF==1.22.1
pyngrok==7.1.6
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.24
pythran==0.10.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scapy==2.4.4
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.8.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
selenium==4.8.3
Send2Trash==1.8.3
service-identity==18.1.0
simplejson==3.14.0
sip @ file:///croot/sip_1698675935381/work
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.5
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.9
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2024.2.12
timm==0.9.2
tinycss2==1.3.0
to==0.3
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
trio==0.24.0
trio-websocket==0.11.1
triton==2.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2023.4
ultralytics==8.1.28
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
utils==1.0.2
uvicorn==0.13.1
virtualenv==20.26.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wsproto==1.2.0
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1696486308932/work
yarl==1.9.4
